refacto(bizdev): refacto side menu cards
